+++ src/backend/core/tests/test_api_config.py
@@ -18,6 +18,8 @@ pytestmark = pytest.mark.django_db
 @override_settings(
     COLLABORATION_WS_URL="http://testcollab/",
     CRISP_WEBSITE_ID="123",
+    FRONTEND_CSS_URL="http://testcss/",
+    FRONTEND_FOOTER_FEATURE_ENABLED=True,
     FRONTEND_THEME="test-theme",
     MEDIA_BASE_URL="http://testserver/",
     POSTHOG_KEY={"id": "132456", "host": "https://eu.i.posthog-test.com"},
@@ -38,6 +40,8 @@ def test_api_config(is_authenticated):
         "COLLABORATION_WS_URL": "http://testcollab/",
         "CRISP_WEBSITE_ID": "123",
         "ENVIRONMENT": "test",
+        "FRONTEND_CSS_URL": "http://testcss/",
+        "FRONTEND_FOOTER_FEATURE_ENABLED": True,
         "FRONTEND_THEME": "test-theme",
         "LANGUAGES": [
             ["en-us", "English"],
@@ -49,4 +53,5 @@ def test_api_config(is_authenticated):
         "MEDIA_BASE_URL": "http://testserver/",
         "POSTHOG_KEY": {"id": "132456", "host": "https://eu.i.posthog-test.com"},
         "SENTRY_DSN": "https://sentry.test/123",
+        "AI_FEATURE_ENABLED": False,
     }