+++ CHANGELOG.md
@@ -10,16 +10,35 @@ and this project adheres to
 
 ## Added
 
+- 🚩(backend) add feature flag for the footer #841
+- 🔧(backend) add view to manage footer json #841
+
+## Changed
+
+- 🚨(frontend) block button when creating doc #749
+
+## Fixed
+
+- 🐛(back) validate document content in serializer #822
+
+## [3.0.0] - 2025-03-28
+
+## Added
+
 - 📄(legal) Require contributors to sign a DCO #779
+- ✨(frontend) add custom css style #771
 
 ## Changed
 
 - ♻️(frontend) Integrate UI kit #783
+- 🏗️(y-provider) manage auth in y-provider app #804
 
 ## Fixed
 
+- 🐛(frontend) conditionally render AI button only when feature is enabled #814
 - 🐛(backend) compute ancestor_links in get_abilities if needed #725
 - 🔒️(back) restrict access to document accesses #801
+
 
 ## [2.6.0] - 2025-03-21
 
@@ -502,8 +521,9 @@ and this project adheres to
 - ✨(frontend) Coming Soon page (#67)
 - 🚀 Impress, project to manage your documents easily and collaboratively.
 
-[unreleased]: https://github.com/numerique-gouv/impress/compare/v2.6.0...main
-[v2.5.0]: https://github.com/numerique-gouv/impress/releases/v2.6.0
+[unreleased]: https://github.com/numerique-gouv/impress/compare/v3.0.0...main
+[v3.0.0]: https://github.com/numerique-gouv/impress/releases/v3.0.0
+[v2.6.0]: https://github.com/numerique-gouv/impress/releases/v2.6.0
 [v2.5.0]: https://github.com/numerique-gouv/impress/releases/v2.5.0
 [v2.4.0]: https://github.com/numerique-gouv/impress/releases/v2.4.0
 [v2.3.0]: https://github.com/numerique-gouv/impress/releases/v2.3.0